Moved services and altered webpack test config
