+++ src/app/core/services/configuration-service/configuration-service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { Http } from '@angular/http';
-import { AppConfig } from './core/models/app-config';
+import { AppConfig } from '../../models/app-config';
 import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/toPromise';
