+++ src/app/features/app/app.spec.ts
@@ -1,6 +1,6 @@
 import { it, describe, expect, beforeEach } from '@angular/core/testing';
 import { AppComponent } from './app.component';
-import { ConfigurationService } from '../../configuration-service';
+import { ConfigurationService } from '../../core/services/configuration-service';
 import { AppConfig } from '../../core/models/app-config';
 
 describe('AppComponent', () => {
