--- src/services/index.ts
@@ -1,2 +1,0 @@
-export { LoggingService } from './logging-service/logging.service';
-export { LoggingErrorHandler } from './logging-service/logging-error-handler';
